SlashCommand refactoring
