+++ Projects/CoX/Servers/MapServer/CMakeLists.txt
@@ -15,7 +15,6 @@ SET(target_CPP
     MapTemplate.cpp
     MessageHelpers.cpp
     NetCommandManager.cpp
-    SlashCommand.cpp
     WorldSimulation.cpp
 )
 SET(target_INCLUDE
@@ -33,7 +32,6 @@ SET(target_INCLUDE
     MapTemplate.h
     MessageHelpers.h
     NetCommandManager.h
-    SlashCommand.h
     WorldSimulation.h
 )
 
@@ -51,8 +49,9 @@ target_link_libraries(MapServer PUBLIC Qt5::Core)
 target_link_libraries(MapServer PUBLIC gameData)
 
 add_subdirectory(ScriptingEngine)
+add_subdirectory(SlashCommands)
 
-target_link_libraries(MapServer PUBLIC Qt5::Core ScriptingEngine)
+target_link_libraries(MapServer PUBLIC Qt5::Core ScriptingEngine SlashCommands)
 
 if(ENABLE_TESTS)
     add_subdirectory(UnitTests)
