+++ Projects/CoX/Servers/MapServer/SlashCommands/SlashCommand_SuperGroup.h
@@ -1,0 +1,16 @@
+/*
+ * SEGS - Super Entity Game Server
+ * http://www.segs.dev/
+ * Copyright (c) 2006 - 2020 SEGS Team (see AUTHORS.md)
+ * This software is licensed under the terms of the 3-clause BSD License. See LICENSE.md for details.
+ */
+
+#pragma once
+
+struct MapClientSession;
+class QStringList;
+
+// Slash Commands related to SuperGroups
+
+// Access Level 9 Commands (GMs)
+void cmdHandler_SetSuperGroup(const QStringList &params, MapClientSession &sess);
